# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Swarm Agent Orchestration API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Agent {
  id: string;
  key: string;
  role: string;
  modelName: string;
  temperature: number;
  description: string;
  maxRetries?: number;
  createdAt: string;
}

export interface AgentInput {
  key: string;
  role: string;
  modelName: string;
  temperature: number;
  description: string;
  maxRetries?: number;
}

export interface AgentUpdate {
  role?: string;
  modelName?: string;
  temperature?: number;
  description?: string;
  maxRetries?: number;
}

export type RunStatus = typeof RunStatus[keyof typeof RunStatus];


export const RunStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export interface Run {
  id: string;
  projectName: string;
  prompt: string;
  status: RunStatus;
  agentKeys: string[];
  parallelCount: number;
  /** @nullable */
  snapshotId?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export type RunDetailStatus = typeof RunDetailStatus[keyof typeof RunDetailStatus];


export const RunDetailStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export type RunLogLevel = typeof RunLogLevel[keyof typeof RunLogLevel];


export const RunLogLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
  think: 'think',
  output: 'output',
} as const;

export interface RunLog {
  id: string;
  runId: string;
  agentKey: string;
  level: RunLogLevel;
  message: string;
  /** @nullable */
  thinkTrace?: string | null;
  createdAt: string;
}

export interface RunDetail {
  id: string;
  projectName: string;
  prompt: string;
  status: RunDetailStatus;
  agentKeys: string[];
  parallelCount: number;
  /** @nullable */
  snapshotId?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  logs: RunLog[];
}

export interface RunInput {
  projectName: string;
  prompt: string;
  agentKeys: string[];
  parallelCount?: number;
  ollamaUrl?: string;
}

export interface Snapshot {
  id: string;
  projectName: string;
  label: string;
  checkpointId: string;
  sizeBytes: number;
  /** @nullable */
  runId?: string | null;
  /** @nullable */
  agentKey?: string | null;
  createdAt: string;
}

export interface SnapshotInput {
  projectName: string;
  label: string;
}

export type WorkspaceFileType = typeof WorkspaceFileType[keyof typeof WorkspaceFileType];


export const WorkspaceFileType = {
  file: 'file',
  directory: 'directory',
} as const;

export interface WorkspaceFile {
  path: string;
  name: string;
  type: WorkspaceFileType;
  /** @nullable */
  size?: number | null;
  /** @nullable */
  children?: WorkspaceFile[] | null;
}

export interface WorkspaceFileContent {
  path: string;
  content: string;
  lines: number;
}

export interface SystemStats {
  totalAgents: number;
  activeRuns: number;
  completedRuns: number;
  failedRuns: number;
  totalSnapshots: number;
  totalLogs: number;
  recentThroughput: number;
}

export type ActivityEntryType = typeof ActivityEntryType[keyof typeof ActivityEntryType];


export const ActivityEntryType = {
  run_started: 'run_started',
  run_completed: 'run_completed',
  run_failed: 'run_failed',
  snapshot_created: 'snapshot_created',
  snapshot_rolled_back: 'snapshot_rolled_back',
  agent_created: 'agent_created',
  agent_updated: 'agent_updated',
} as const;

export type ActivityEntryEntityType = typeof ActivityEntryEntityType[keyof typeof ActivityEntryEntityType];


export const ActivityEntryEntityType = {
  run: 'run',
  snapshot: 'snapshot',
  agent: 'agent',
} as const;

export interface ActivityEntry {
  id: string;
  type: ActivityEntryType;
  message: string;
  entityId: string;
  entityType: ActivityEntryEntityType;
  createdAt: string;
}

export type ListRunsParams = {
status?: ListRunsStatus;
limit?: number;
};

export type ListRunsStatus = typeof ListRunsStatus[keyof typeof ListRunsStatus];


export const ListRunsStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export type ListWorkspaceFilesParams = {
project?: string;
};

export type GetWorkspaceFileParams = {
path: string;
};

export type ListActivityParams = {
limit?: number;
};