# lib/api-zod/src/generated/types/activityEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Swarm Agent Orchestration API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEntryEntityType } from './activityEntryEntityType';
import type { ActivityEntryType } from './activityEntryType';

export interface ActivityEntry {
  id: string;
  type: ActivityEntryType;
  message: string;
  entityId: string;
  entityType: ActivityEntryEntityType;
  createdAt: string;
}