# lib/api-zod/src/generated/types/agentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Swarm Agent Orchestration API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentInput {
  key: string;
  role: string;
  modelName: string;
  temperature: number;
  description: string;
  maxRetries?: number;
}