# lib/api-zod/src/generated/types/agentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Swarm Agent Orchestration API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentUpdate {
  role?: string;
  modelName?: string;
  temperature?: number;
  description?: string;
  maxRetries?: number;
}