# lib/api-zod/src/generated/types/snapshotInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Swarm Agent Orchestration API
 * OpenAPI spec version: 0.1.0
 */

export interface SnapshotInput {
  projectName: string;
  label: string;
}